% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{add_colors}
\alias{add_colors}
\title{If colors arenot defined, add colors from Color Brewer, up to 8}
\usage{
add_colors(df)
}
\arguments{
\item{df}{Dataframe with a column called \code{name} that has a unique name for each dataset}
}
\value{
A dataframe with an extra column called \code{color} that has a unique color for each dataset
}
\description{
If colors arenot defined, add colors from Color Brewer, up to 8
}
