% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carpenter_williams_1995}
\alias{carpenter_williams_1995}
\title{Digitised data corresponding to Figure 2 of Carpenter and Williams (1995)}
\format{
\subsection{\code{carpenter_williams_1995}}{

A dataframe of 20014 reaction times for participant a and 22518 reaction times for participant b
\describe{
\item{participant}{Participant "name", either "a" or "b"}
\item{condition}{Prior percentage probability of the target being in the location of the eye movement (p05, p10, p25, p50, p75, p90, or p95)}
\item{time}{Saccadic latency in ms}
}
}
}
\source{
Carpenter, R. H., & Williams, M. L. L. (1995). Neural computation of log likelihood in control of saccadic eye movements. Nature, 377(6544), 59-62.
}
\usage{
carpenter_williams_1995
}
\description{
Digitised data corresponding to Figure 2 of Carpenter and Williams (1995)
}
\keyword{datasets}
