% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{convert_to_seconds}
\alias{convert_to_seconds}
\title{Convert reaction time vector to seconds. Used only internally, so not exported}
\usage{
convert_to_seconds(rt_vector, time_units = "ms")
}
\arguments{
\item{rt_vector}{Vector of reaction times, with type 'double'}

\item{time_units}{Units of rt_vector, must be one of "ms", "ds", or "s"}
}
\value{
rt_vector in seconds
}
\description{
Convert reaction time vector to seconds. Used only internally, so not exported
}
