% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{individual_later_fit}
\alias{individual_later_fit}
\title{Fit individual LATER model to each dataset in a dataframe of datasets}
\usage{
individual_later_fit(df, with_early_component = FALSE)
}
\arguments{
\item{df}{A dataframe with columns: \code{time}, \code{name}, \code{promptness}, and \code{e_cdf}}

\item{with_early_component}{If \code{TRUE}, the model contains a second 'early'
component that is absent when \code{FALSE} (the default).}
}
\value{
A dataframe with one row for each named dataset in \code{df} and columns
equal to the LATER model parameters returned by fit_data$named_fit_params
}
\description{
Fit individual LATER model to each dataset in a dataframe of datasets
}
\examples{
a <- dplyr::filter(carpenter_williams_1995, participant == "a")
df <- prepare_data(a)
fit_params <- individual_later_fit(df)
}
