% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{model_cdf}
\alias{model_cdf}
\title{Evalulate the cumulative distribution function under the model.}
\usage{
model_cdf(q, later_mu, later_sd, early_sd = NULL)
}
\arguments{
\item{q}{Vector of quantiles}

\item{later_mu}{Vector of the mean of the later component.}

\item{later_sd}{Vector of the standard deviation of the later component.}

\item{early_sd}{Vector of the standard deviation of the early component,
or \code{NULL} if there is no early component (the default).}
}
\value{
Vector of cumulative distribution values
}
\description{
Evalulate the cumulative distribution function under the model.
}
\examples{
model_cdf(q = 1, later_mu = 1, later_sd = 1)
model_cdf(q = 1, later_mu = 1, later_sd = 1, early_sd = 3)
}
