% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprobit_plot.R
\name{reciprobit_plot}
\alias{reciprobit_plot}
\title{Plot reaction times in reciprobit axes}
\usage{
reciprobit_plot(plot_data)
}
\arguments{
\item{plot_data}{A dataframe with columns: \code{time}, \code{color}, \code{name}, \code{promptness},
and \code{e_cdf}; one color per name (participant)}
}
\value{
A reciprobit plot with the cumulative probability distribution of the reaction times
}
\description{
Plot reaction times in reciprobit axes
}
\examples{
df <- prepare_data(rt2)
reciprobit_plot(df)
}
