% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reddi_asrress_carpenter_2003}
\alias{reddi_asrress_carpenter_2003}
\title{Digitised data corresponding to Figure 2 of Reddi, Asrress, and Carpenter,
2003.}
\format{
\subsection{\code{reddi_asrress_carpenter_2003}}{

A dataframe of 1600 reaction times for participant J
\describe{
\item{participant}{Participant "name", equal to "J"}
\item{condition}{Different degrees of coherence of the kinematograms shown
as stimuli, (64\%, 32\%, 16\%, 8\%: S, low-urgency–slow and
accurate–conditions)}
\item{time}{Saccadic latency in ms}
}
}
}
\source{
Reddi, B. A., Asrress, K. N., & Carpenter, R. H. (2003). Accuracy,
information, and response time in a saccadic decision task. Journal of
Neurophysiology, 90(5), 3538-3546.
}
\usage{
reddi_asrress_carpenter_2003
}
\description{
Digitised data corresponding to Figure 2 of Reddi, Asrress, and Carpenter,
2003.
}
\keyword{datasets}
