% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rt}
\alias{rt}
\title{LATER unit reaction time data}
\format{
\subsection{\code{rt}}{

A vector of 200 reaction times
\describe{
\item{rt}{Vector of reaction times in ms}
}
}
}
\source{
LATER 1 mu 5 sig 0.5 R, SPIC program, developed by R.H.S. Carpenter
}
\usage{
rt
}
\description{
Simulated data from SPIC
}
\keyword{datasets}
