% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rt2}
\alias{rt2}
\title{Two LATER units reaction time data}
\format{
\subsection{\code{rt2}}{

A dataframe of 200 reaction times for two participants: mu3 and mu4
\describe{
\item{time}{Reaction times in ms}
\item{name}{Participant name}
\item{color}{Color to use for plotting each participant}
}
}
}
\source{
mu3 = LATER 1 mu 3 sig 0.5 R, count = 200; mu4 = LATER 2 mu 4 sig 0.5 R, count = 200 - SPIC program, developed by R.H.S. Carpenter
}
\usage{
rt2
}
\description{
Simulated data from SPIC
}
\keyword{datasets}
