% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprobit_plot.R
\name{reciprobit_plot}
\alias{reciprobit_plot}
\title{Plot reaction times in reciprobit axes}
\usage{
reciprobit_plot(rt)
}
\arguments{
\item{rt}{vector of reaction times}
}
\value{
A reciprobit plot with the cumulative probability distribution of the reaction times
}
\description{
Plot reaction times in reciprobit axes
}
\examples{
reciprobit_plot(rt)
}
